The source map is unnecessary for production build.
It slows down the build.

Benchmark:

| Before | After |
| :-----: | :----: |
| 1m5.159s | 0m54.477s |

`Built on a VPS with 4 virtual cores and 8GB memory.`

And parallelization makes it a lot faster.

Benchmark:

| Before | After |
| :-----: | :----: |
| 1m5.159s | 0m43.487s |

`Built on a VPS with 4 virtual cores and 8GB memory.`