@imbsky 2018-10-06
`mangle` is also one cause of slow down build speed, but I really don't know if it's necessary.
Can I get your opinion?