@ykzts 2018-10-06
> The source map is unnecessary for production build.

Mastodon intentionally outputs source map in production.