@ykzts 2018-10-06
I think that `mangle` can be changed to` false`.