@nightpool 2018-10-06
What is the bundle size before and after this commit? I feel like this change was made too hastily—mangle is a key compression tool, and it would be bad if this change increased the bundle download time for all users just to make compilation slightly faster for admins.