@nightpool 2018-10-06
I can't find any difference in bundle sizes when I was testing this locally, but I suspect that means that the mangle option *wasn't* turned off (because it's `true` by default according to https://github.com/mishoo/UglifyJS2#minify-options)